支持使用 clang-cl 编译 (#1195)
* chore: native 项目支持 clang-cl

* chore: 修复 clang 编译错误

* chore: 修复部分 clang 编译警告

* chore: 修复警告

* chore: 修复所有警告，优化 llvm 查找

* chore: 启用 LTO

* chore: 支持 ARM64

* chore: _ConanDeps 支持 clang-cl

* chore: 编译选项改变 _ConanDeps 自动重新编译

* chore: profile 更改时重新编译 _ConanDeps

* chore: 将 hu 和 ka 加入项目文件，但不参与生成
否则不参与批量替换

* chore: Magpie 项目支持并行编译

* chore: 添加几个提高性能的编译选项

* chore: 优化 _ConanDeps

* chore: publish.py 支持参数

* chore: CI 支持 clang 编译

* chore: 修复 CI 的 conan 缓存

* chore: 优化编译选项，修复 CI

* chore: 修复 CI

* chore: 改变参数顺序

* chore: 添加编译选项支持针对当前硬件生成优化代码

* chore: 优化 CI 脚本

* chore: publish.py 添加 --use-native-march 选项

* chore: 脚本集中在 scripts 文件夹，msbuild 启用并行编译

* chore: clang, x64 配置启用 CX16 指令

* chore: 更新依赖

* chore: 更新格式设置

* chore: 增加额外的斜杠以提高兼容性

* chore: 修复效果文件复制

+++ src/Magpie/AppSettings.cpp
@@ -1,22 +1,22 @@
 #include "pch.h"
+#include "App.h"
 #include "AppSettings.h"
+#include "AutoStartHelper.h"
+#include "CommonSharedConstants.h"
+#include "JsonHelper.h"
+#include "LocalizationService.h"
+#include "Logger.h"
+#include "MainWindow.h"
+#include "Profile.h"
+#include "resource.h"
+#include "ScalingMode.h"
+#include "ScalingModesService.h"
+#include "ShortcutHelper.h"
 #include "StrHelper.h"
 #include "Win32Helper.h"
-#include "Logger.h"
-#include "ShortcutHelper.h"
-#include "Profile.h"
-#include "CommonSharedConstants.h"
-#include "AutoStartHelper.h"
-#include "ScalingModesService.h"
-#include "JsonHelper.h"
-#include "ScalingMode.h"
-#include "LocalizationService.h"
-#include "resource.h"
-#include "App.h"
-#include "MainWindow.h"
 #include <rapidjson/prettywriter.h>
-#include <ShlObj.h>
 #include <ShellScalingApi.h>
+#include <ShlObj.h>
 
 using namespace winrt;
 using namespace winrt::Magpie;
@@ -950,7 +950,7 @@ bool AppSettings::_LoadProfile(
 	}
 
 	JsonHelper::ReadInt(profileObj, "scalingMode", profile.scalingMode);
-	if (profile.scalingMode < -1 || profile.scalingMode >= _scalingModes.size()) {
+	if (profile.scalingMode < -1 || profile.scalingMode >= (int)_scalingModes.size()) {
 		profile.scalingMode = -1;
 	}
